+++ src/main/src/processEntry/mapper/getAssetFields.test.ts
