+++ src/main/src/processEntry/mapper/getAssetFields.ts
