+++ src/main/src/processEntry/mapper/getEntryFields.test.ts
