+++ src/main/src/processEntry/mapper/getEntryFields.ts
