+++ src/main/src/processEntry/mapper/index.ts
@@ -2,11 +2,11 @@
 import { documentToPlainTextString as richTextToPlain } from '@contentful/rich-text-plain-text-renderer';
 import { Asset, Entry } from 'contentful';
 import { ResolveEntryConfig } from '@main/src/config';
-import getEntryFields from './src/getEntryFields';
-import getAssetFields from './src/getAssetFields';
-import richTextToMarkdown from './src/richTextToMarkdown';
-import richTextNodes from './src/richTextNodes';
-import { OverrideConfig } from '../config/src/types';
+import getEntryFields from './getEntryFields';
+import getAssetFields from './getAssetFields';
+import richTextToMarkdown from './richTextToMarkdown';
+import richTextNodes from './richTextNodes';
+import { OverrideConfig } from '../../config/src/types';
 
 const mapArrayField = (
     fieldContent: Entry<any>[] | Asset[] | string[]
