+++ src/main/src/processEntry/mapper/richTextNodes.ts
