+++ src/main/src/processEntry/mapper/richTextToMarkdown.ts
